+++ tests/tada_commands_tests.ads
@@ -7,6 +7,8 @@ package Tada_Commands_Tests is
 
    procedure Test_Parse_Version (T : in out Test_Context);
 
+   procedure Test_Parse_Doc (T : in out Test_Context);
+
    procedure Test_Parse_Build (T : in out Test_Context);
 
    procedure Test_Parse_Build_Debug (T : in out Test_Context);